refactor network_share.py, system.py and backup_preparator.py
