+++ base/logic/network_share.py
@@ -20,24 +20,32 @@ class NetworkShare:
         return self._available
 
     def mount_datasource_via_smb(self) -> None:
+        self._create_mountpoint()
+        self._perform_mount()
+        self._available = HddState.available
+
+    def unmount_datasource_via_smb(self) -> None:
+        self._perform_unmount()
+        self._available = HddState.not_available
+
+    def _create_mountpoint(self) -> None:
         try:
             Path(self._config.local_nas_hdd_mount_point).mkdir(exist_ok=True)
         except FileExistsError:
             pass  # exist_ok=True was intended to supress this error, however it works in a different way
         except OSError as e:
             LOG.warning(f"strange OS-Error occured on trying to create the NAS-HDD Mountpoint: {e}")
+
+    def _perform_mount(self) -> None:
         try:
             SmbShareMount().mount_smb_share(self._config.local_nas_hdd_mount_point)
         except NetworkError as e:
             self._available = HddState.unknown
             raise NetworkError from e
-        self._available = HddState.available
 
-    def unmount_datasource_via_smb(self) -> None:
+    def _perform_unmount(self) -> None:
         try:
             SmbShareMount().unmount_smb_share(self._config.local_nas_hdd_mount_point)
         except NetworkError as e:
             self._available = HddState.unknown
             raise NetworkError from e
-        else:
-            self._available = HddState.not_available
